+++ Tests/Controller/OrmResourceControllerTest.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Lemon\RestBundle\Tests\Controller;
-
-class OrmResourceControllerTest extends ResourceControllerTest
-{
-}