+++ Tests/Fixtures/Tag.php
@@ -2,14 +2,12 @@
 namespace Lemon\RestBundle\Tests\Fixtures;
 
 use Doctrine\ORM\Mapping as ORM;
-use Doctrine\ODM\MongoDB\Mapping\Annotations as MongoDB;
 use Symfony\Component\Validator\Constraints as Assert;
 use JMS\Serializer\Annotation as Serializer;
 
 /**
  * @ORM\Table()
  * @ORM\Entity()
- * @MongoDB\Document
  */
 class Tag
 {
@@ -17,7 +15,6 @@ class Tag
      * @ORM\Column(name="id", type="integer", nullable=false)
      * @ORM\Id
      * @ORM\GeneratedValue(strategy="IDENTITY")
-     * @MongoDB\Id
      * @Serializer\Type("string")
      */
     public $id;
@@ -25,7 +22,6 @@ class Tag
     /**
      * @ORM\Column(name="name", type="string", length=255, nullable=false)
      * @Assert\NotBlank()
-     * @MongoDB\Field(type="string")
      */
     public $name;
 }